Update Ukrainian translation
Hi, this is a updated Ukrainian translation (include interface and documentation). However, unfortunately, some strings are not available for translation. For example, messages about successful video downloads or the download progress bar are spoken in the original language of the addon and cannot be translated. Please fix this issue and let us know if it is possible and include this translation in the addon.